Create StringCalculation
Get an input string S from the user.  If the string consists of ‘*’ and ‘#’, program needs to find the minimum number of ‘*’ or ‘#’ to make it a valid string. The string is considered valid if the number of ‘*’ and ‘#’ are equal. The ‘*’ and ‘#’ can be at any position in the string.
Note : The output will be zero, positive or negative integer based on number of ‘*’ and ‘#’ in the input string.